Move `update_opening_hours_hash` to actions class
